+++ src/app/app.rs
@@ -1,7 +1,7 @@
 use std::sync::Arc;
 
 use crate::app::ports::{OAuthService, PushService, SupportersService, TokenRepository};
-use crate::push::client::PushClient;
+use crate::push::infra::PushClient;
 use crate::token_manager::application::TokenManagerS;
 
 pub struct App {
