+++ src/app/mod.rs
@@ -1,3 +1,5 @@
 pub mod app;
 pub mod ports;
 pub mod service_adapters;
+
+pub use self::app::{App, AppBuilder};
