+++ src/push/mod.rs
@@ -1,4 +1,3 @@
 pub mod api;
-pub mod client;
 pub mod domain;
 pub mod infra;
